+++ HumanResourceManagementSystem/src/main/java/kodlamaio/HumanResourceManagementSystem/business/abstracts/CandidateService.java
@@ -12,13 +12,13 @@ import java.util.List;
 
 public interface CandidateService {
     Result add(CandidateAddDto candidateAddDto);
-    Result update(int id,CandidateUpdateDto candidateUpdateDto);
-    Result delete(int id);
+    Result update(Long id,CandidateUpdateDto candidateUpdateDto);
+    Result delete(Long id);
     Result updateTest(Candidate candidate);
 
 
 
     DataResult<List<Candidate>> getAll();
-    DataResult<Candidate> getById(int id);
+    DataResult<Candidate> getById(Long id);
     DataResult<Candidate> getByNationalIdentityNumber(String nationalIdentityNumber);
 }
