+++ HumanResourceManagementSystem/src/main/java/kodlamaio/HumanResourceManagementSystem/business/abstracts/CovertLetterService.java
@@ -10,10 +10,10 @@ import javax.xml.crypto.Data;
 import java.util.List;
 
 public interface CovertLetterService {
-    Result add(CoverLetterDto coverLetterDto,int cvId);
-    Result delete(int cvId,int coverLetterId);
-    Result update(CoverLetterDto coverLetterDto,int cvId,int coverLetterId);
+    Result add(CoverLetterDto coverLetterDto,Long cvId);
+    Result delete(Long cvId,Long coverLetterId);
+    Result update(CoverLetterDto coverLetterDto,Long cvId,Long coverLetterId);
 
-    DataResult<List<CoverLetter>> getAll(int cvId);
-    DataResult<CoverLetter> getOne(int id);
+    DataResult<List<CoverLetter>> getAll(Long cvId);
+    DataResult<CoverLetter> getOne(Long id);
 }
